fix(env-check): added volume config parameters for RunJob utility
